Use line-height for navbar items spacing
Because the navbar has an 'a' element, using gap or padding makes it so
only the actual link text may be clicked. This also made styling via
:hover misleading, because highlighting the padding did not mean that
you were on top of a clickable element.

Using line-height changes the actual link to take up more vertical space
which helps with clicking and styling.

Block level padding and grid gaps were removed from the sidebar for this
reason.